refactor: combine Sound and SoundKt
